# openapi/account-api.yaml
openapi: 3.1.0
info:
  title: Account Center Management API
  version: 1.0.0
  description: >-
    Account 管理端同源 Session API。浏览器不得持有应用 Secret、Session ID 或长期 Token。
    管理端 message/fieldErrors 根据可选 Accept-Language 返回中文或英文；支持 zh/en 地区变体与权重，
    缺失、不支持或非法头回退 zh-CN。响应包含 Content-Language 和 Vary: Accept-Language。
    错误码、字段名、状态值及用户业务数据不随语言变化。
servers:
  - url: /
...
paths:
...
  /api/users:
    get:
      tags: [Users]
      operationId: listUsers
      summary: 分页查询全局用户
      security:
        - sessionCookie: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: query
          in: query
          description: 匹配账号、姓名、邮箱或手机号
          schema:
            type: string
            maxLength: 128
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/UserStatus'
        - name: sort
          in: query
          schema:
            type: string
            enum:
              - createdAt,asc
              - createdAt,desc
              - account,asc
              - account,desc
              - name,asc
              - name,desc
            default: createdAt,desc
      responses:
        '200':
          description: 稳定分页结果；相同排序值始终以 id 升序作为次级排序
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationFailed'
    post:
      tags: [Users]
      operationId: createUser
      summary: 创建全局用户
      security:
        - sessionCookie: []
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: 用户已创建；重复幂等请求复用首次响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /api/users/{userId}:
    get:
      tags: [Users]
      operationId: getUser
      summary: 查询全局用户详情
      security:
        - sessionCookie: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 64
      responses:
        '200':
          description: 用户详情；不含密码、角色、应用 Secret 和内部审计字段
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Users]
      operationId: updateUser
      summary: 修改全局用户基本资料
      security:
        - sessionCookie: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 64
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: 修改后的用户；版本单调递增
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /api/users/{userId}/status:
    put:
      tags: [Users]
      operationId: changeUserStatus
      summary: 启用或禁用全局用户
      security:
        - sessionCookie: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 64
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeUserStatusRequest'
      responses:
        '200':
          description: 状态已更新；不会在本切片触发业务系统同步
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
  parameters:
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 32
      description: 由 GET /api/session 初始化并只保存在前端内存中的同步器 Token
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
      description: 按操作者、方法和规范化资源路径隔离，完成结果至少保留 24 小时
...
  responses:
    Unauthorized:
      description: 未登录或 Session 已失效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: CSRF 校验失败或当前角色没有访问能力
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationFailed:
      description: 请求字段校验失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: 账号重复、资源版本冲突或幂等键冲突
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    UserStatus:
      type: string
      enum: [enabled, disabled]
    UserResponse:
      type: object
      additionalProperties: false
      required: [id, account, email, name, phone, status, version]
      properties:
        id:
          type: string
          maxLength: 64
        account:
          type: string
        email:
          type: string
        name:
          type: string
        phone:
          type: string
        status:
          $ref: '#/components/schemas/UserStatus'
        version:
          type: integer
          format: int64
          minimum: 1
        createdAt:
          type: [string, 'null']
          format: date-time
        updatedAt:
          type: [string, 'null']
          format: date-time
    CreateUserRequest:
      type: object
      additionalProperties: false
      required: [account, email, name, phone]
      properties:
        account:
          type: string
          minLength: 1
          maxLength: 128
        email:
          type: string
          format: email
          minLength: 1
          maxLength: 255
        name:
          type: string
          minLength: 1
          maxLength: 128
        phone:
          type: string
          minLength: 1
          maxLength: 64
    UpdateUserRequest:
      type: object
      additionalProperties: false
      required: [account, email, name, phone, version]
      properties:
        account:
          type: string
          minLength: 1
          maxLength: 128
        email:
          type: string
          format: email
          minLength: 1
          maxLength: 255
        name:
          type: string
          minLength: 1
          maxLength: 128
        phone:
          type: string
          minLength: 1
          maxLength: 64
        version:
          type: integer
          format: int64
          minimum: 1
    ChangeUserStatusRequest:
      type: object
      additionalProperties: false
      required: [status, version, reason]
      properties:
        status:
          $ref: '#/components/schemas/UserStatus'
        version:
          type: integer
          format: int64
          minimum: 1
        reason:
          type: string
          minLength: 1
          maxLength: 256
    UserPageResponse:
      type: object
      additionalProperties: false
      required: [items, page, size, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        page:
          type: integer
          minimum: 1
        size:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          format: int64
          minimum: 0
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message, traceId]
      properties:
        code:
          type: string
          examples: [AUTHENTICATION_REQUIRED]
        message:
          type: string
        traceId:
          type: string
        fieldErrors:
          type: object
          additionalProperties:
            type: string
...